0912/ch1 complete

+++ src/ch1/Q9.java
@@ -3,8 +3,6 @@ package ch1;
 public class Q9 {
     public static void main(String[] args) {
         int n=2000000000;
-        System.out.println((long)n+n);
-//        System.out.println((long)n+(long)n);
-
+        System.out.println((long)n+(long)n);
     }
 }
